不正なスキーマファイルを修正 + 出力先のmarkdownファイルのパスの生成処理をschema_id_util.pyに移動
