+++ schema/objects/transactions/transfer/StockTransfer.schema.json
@@ -1,5 +1,5 @@
 {
-  "id": "https://jocf.startupstandard.org/jocf/main/schema/objects/transactions/transfer/StockTransfer.schema.json",
+  "$id": "https://jocf.startupstandard.org/jocf/main/schema/objects/transactions/transfer/StockTransfer.schema.json",
   "title": "株式異動トランザクション",
   "description": "株式の所有権移転や売買を表現するトランザクション",
   "type": "object",
