+++ schema/types/File.schema.json
@@ -10,7 +10,7 @@
     },
     "md5": {
       "description": "MD5ファイルチェックサム",
-      "$ref": "https://jocf.startupstandard.org/jocf/schema/types/Md5.schema.json"
+      "$ref": "https://jocf.startupstandard.org/jocf/main/schema/types/Md5.schema.json"
     }
   },
   "required": ["filepath", "md5"]
